ssafy-backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
beautifulsoup4==4.8.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.41.0
django-bootstrap4==1.0.1
django-cors-headers==3.2.1
django-drf-filepond==0.2.2
django-extensions==2.2.1
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
Faker==2.0.2
firebase-admin==3.2.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
httplib2==0.16.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
npm==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.9
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
shortuuid==0.5.0
six==1.14.0
soupsieve==1.9.3
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.7
wrapt==1.11.2
